chore: add patch files for serde support in comrak-labs
Remarks: I decided to take a different approach for the full rewrite of
comrak-wasm when I discovered that comrak did not support serde features out of
the box. Instead of spending a bunch of time duplicating comrak's options
interfaces and AST structures in comrak-wasm (just to add serde support), I
opted instead to add serde support directly into comrak itself, allowing its
data structures to be serialized and deserialized directly into JavaScript
objects via serde-wasm-bindgen.

Things got a bit more complicated when it came to implementing serialization
for the AST structures, since comrak uses an arena allocator and cyclic node
references. While not the perfect solution (and certainly not the prettiest), I
ended up writing a custom serde impl that leverages integer indices and a
single vector of nodes at the root. This allows an AST to be serialized AND
deserialized correctly, without any data loss or problems with circular
references, while still maintaining the integrity of the arena allocation
system.

The one gotcha here is that the serialized AST is flat in structure. In
JS-land, it will need to be further processed (relinked) to reconstruct the
tree structure.

Since these are extremely bleeding-edge features, I opted to fork comrak rather
than submitting these changes as a PR to the main comrak repo just yet. I'd
love to get @kivikakk's input on this and whether or not serde support is
something that would fit into comrak's future plans.